requirements.txt
python-dotenv>=1.0.0
langgraph>=0.2.45
langchain>=0.3.0
langchain-core>=0.3.0
langchain-ollama>=0.2.0
langchain-openai>=0.3.0
chainlit>=2.8.3
pydantic>=2.0.0
typing-extensions>=4.0.0
pytest>=8.0.0
